Worked on behaviors and enemy states
Could not figure out how to pull the changes Joshua made (seriously pull origin just wasn't working?), so I just copied and pasted what he changed from the github cloud. Mostly just worked on Behavior.java and making it so it's actually closer to being correct.

Unable to "make all" because of the playerUnits call in Behavior. Not sure if I need to add playerUnits as a method parameter or need to use a different method. But that's the progress yeah